kanban_backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
daphne
whitenoise
gunicorn
